fs can't be used on client-side
