+++ src/pages/sitemap.xml.js
@@ -1,7 +1,5 @@
-import fs from 'fs';
-import { join } from 'path';
-
-import { getAllPosts, getAllExternalReferences } from '../lib/blog';
+import { getStaticPageUrls } from '../lib/staticPagesLoader';
+import { getAllPosts } from '../lib/blog';
 import { getAllTags } from '../lib/tags';
 import config from '../config';
 
@@ -17,21 +15,7 @@ const urlEntryForPage = (url, frequency = 'daily') => `
 `;
 
 export const getServerSideProps = async ({ res }) => {
-  const staticPagesDirectory = join(process.cwd(), 'src', 'pages');
-  const staticPages = fs
-    .readdirSync(staticPagesDirectory)
-    .filter((staticPage) => {
-      return ![
-        '[slug].js',
-        '_app.js',
-        '_document.js',
-        '_error.js',
-        'sitemap.xml.js',
-      ].includes(staticPage);
-    })
-    .map((staticPagePath) => {
-      return `${siteUrl}/${staticPagePath.replace('.js', '')}`;
-    });
+  const staticPageUrls = getStaticPageUrls();
 
   const allPostUrls = (await getAllPosts()).map(
     (post) => `${siteUrl}/${post.frontmatter.path}`
@@ -41,11 +25,9 @@ export const getServerSideProps = async ({ res }) => {
     (tag) => `${siteUrl}/tags/${tag}`
   );
 
-  console.log(allPostUrls);
-
   const sitemap = `<?xml version="1.0" encoding="UTF-8"?>
     <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-      ${staticPages.map((page) => urlEntryForPage(page)).join('')}
+      ${staticPageUrls.map((page) => urlEntryForPage(page)).join('')}
       ${allPostUrls.map((post) => urlEntryForPage(post)).join('')}
       ${allTagUrls.map((tag) => urlEntryForPage(tag)).join('')}
     </urlset>
